# openapi.yaml
openapi: 3.0.3
info:
  title: HM Software API
  version: 0.1.0
  description: Phase 0 (OPD) + Phase 1 (Lab) modular monolith backend
paths:
...
  /api/v1/orders/:
    get:
      operationId: v1_orders_list
      description: |-
        Thin API layer:
        - scope parsing
        - idempotency caching
        - serializers validation
        - delegates writes to OrderService, reads to OrderSelector
      parameters:
      - in: header
        name: X-Facility-Id
        schema:
          type: string
        required: true
      - in: header
        name: X-Tenant-Id
        schema:
          type: string
        required: true
      tags:
      - Orders
      security:
      - CookieOrBearerJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
          description: ''
    post:
      operationId: v1_orders_create
      description: |-
        Thin API layer:
        - scope parsing
        - idempotency caching
        - serializers validation
        - delegates writes to OrderService, reads to OrderSelector
      tags:
      - v1
      security:
      - CookieOrBearerJWT: []
      responses:
        '201':
          description: No response body
  /api/v1/orders/{id}/:
    get:
      operationId: v1_orders_retrieve
      description: |-
        Thin API layer:
        - scope parsing
        - idempotency caching
        - serializers validation
        - delegates writes to OrderService, reads to OrderSelector
      parameters:
      - in: header
        name: X-Facility-Id
        schema:
          type: string
        required: true
      - in: header
        name: X-Tenant-Id
        schema:
          type: string
        required: true
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - Orders
      security:
      - CookieOrBearerJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: ''
...
  /api/v1/tenants/:
    get:
      operationId: v1_tenants_retrieve
      description: |-
        Admin-only tenant management.
        Routing is centralized in hm_core/api/urls.py (locked standard).
      tags:
      - v1
      security:
      - CookieOrBearerJWT: []
      responses:
        '200':
          description: No response body
    post:
      operationId: v1_tenants_create
      description: |-
        Admin-only tenant management.
        Routing is centralized in hm_core/api/urls.py (locked standard).
      tags:
      - v1
      security:
      - CookieOrBearerJWT: []
      responses:
        '201':
          description: No response body
  /api/v1/tenants/{id}/:
    get:
      operationId: v1_tenants_retrieve_2
      description: |-
        Admin-only tenant management.
        Routing is centralized in hm_core/api/urls.py (locked standard).
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - CookieOrBearerJWT: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        encounter:
          type: string
          format: uuid
        order_type:
          $ref: '#/components/schemas/OrderTypeEnum'
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        status:
          $ref: '#/components/schemas/OrderStatusEnum'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          readOnly: true
      required:
      - encounter
      - id
      - items
      - order_type
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        service_code:
          type: string
          maxLength: 64
          pattern: ^[-a-zA-Z0-9_]+$
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        status:
          $ref: '#/components/schemas/OrderItemStatusEnum'
      required:
      - id
      - service_code
    OrderItemStatusEnum:
      enum:
      - REQUESTED
      type: string
      description: '* `REQUESTED` - Requested'
    OrderStatusEnum:
      enum:
      - CREATED
      - CANCELLED
      type: string
      description: |-
        * `CREATED` - Created
        * `CANCELLED` - Cancelled
    OrderTypeEnum:
      enum:
      - LAB
      type: string
      description: '* `LAB` - Lab'
...
    PriorityEnum:
      enum:
      - ROUTINE
      - URGENT
      - STAT
      type: string
      description: |-
        * `ROUTINE` - Routine
        * `URGENT` - Urgent
        * `STAT` - Stat
...
  securitySchemes:
    CookieOrBearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Send `Authorization: Bearer <access>` OR rely on HttpOnly cookie
        (e.g. `hm_access`). Scope headers required: `X-Tenant-ID`, `X-Facility-ID`.'